+++ app/src/main/java/com/codesoom/assignment/advice/NotFoundErrorAdvice.java
@@ -1,7 +1,7 @@
 package com.codesoom.assignment.advice;
 
 import com.codesoom.assignment.services.product.exception.ProductNotFoundException;
-import com.codesoom.assignment.UserNotFoundException;
+import com.codesoom.assignment.services.user.exception.UserNotFoundException;
 import com.codesoom.assignment.advice.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
