+++ app/src/main/java/com/codesoom/assignment/services/user/UserUpdateService.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.services.user;
 
-import com.codesoom.assignment.UserNotFoundException;
+import com.codesoom.assignment.services.user.exception.UserNotFoundException;
 import com.codesoom.assignment.services.user.domain.User;
 import com.codesoom.assignment.services.user.infra.JpaUserRepository;
 import org.springframework.stereotype.Service;
