+++ app/src/main/java/com/codesoom/assignment/services/user/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment;
+package com.codesoom.assignment.services.user.exception;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(Long id) {
